+++ pkg/wallet/api/ethgrp/eth/util.go
@@ -127,5 +127,5 @@ func (e *Ethereum) FromGWei(v int64) *big.Int {
 
 // FromFloatEther converts Ether(float64) to Wei(*big.Int)
 func (e *Ethereum) FromFloatEther(v float64) *big.Int {
-	return big.NewInt(int64(v * float64(params.Ether)))
+	return big.NewInt(int64(v * params.Ether))
 }
